Cjl/litt integration (#3435)
## Describe your changes and provide context

- Make it so that LittDB is included in the regular build system. 
- Fixed all of the lint errors exposed by integrating the build.
- Causes CI to run LittDB tests. 
- sabled some tests that were pulling in a docker dependency. The tests
were slow, mildly flaky, and the features they covered are not important
to our use case.

## Testing performed to validate your change

unit test coverage, not yet used in prod

---------

+++ sei-db/db_engine/litt/test/unlock_test.go
@@ -1,5 +1,3 @@
-//go:build littdb_wip
-
 package test
 
 import (
@@ -26,7 +24,7 @@ func TestUnlock(t *testing.T) {
 	config, err := litt.DefaultConfig(volumes...)
 	config.Fsync = false // Disable fsync for faster tests
 	config.TargetSegmentFileSize = 100
-	config.ShardingFactor = uint32(len(volumes))
+	config.ShardingFactor = uint8(len(volumes))
 	require.NoError(t, err)
 
 	db, err := littbuilder.NewDB(config)
